Version 2 (5/8/2019)
uses gui to read files and the table name. Minimal error checking. The csv file is read and the table is created in the Sqlite db. A new db is created if the one specified does not exist.

After the conversion, the first 20 lines of the csv file and the table are displayed for review and error checking.