Version 2 (5/8/2019)
uses gui to read files and the table name. Minimal error checking. The csv file is read and the table is created in the Sqlite db. A new db is created if the one specified does not exist.

After the conversion, the first 20 lines of the csv file and the table are displayed for review and error checking.

Version 3

Break up the convert() function into smaller functions.
Add a listbox to show the column names.
Added more error checking
Added ability to enter a new table name if the one entered already exists
fixed error checking and messages for table not created
Added fields on the form for updating the column heading and column type

Version 5/12/2019
Configured popups to stay on top
went back to original code for removing commas from "real" and "int" columns